requirements.txt
supabase>=2.0.0
httpx>=0.25.0
crawl4ai>=0.3.0
gspread>=5.12.0
google-auth>=2.23.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
